Add ConcurrentBoardState class
